Adding alternative instructions
I didn't want to install `poetry` just for this project that I intend to use only once. So I came up with a different set of instructions.

Also… In the end I haven't managed to run the project, because it requires `google-chrome`, which I don't have on my headless Raspberry Pi that doesn't even has X installed.

Also also… I'm not sure if it is Alternate or Alternative.